+++ spec/examples/disallow_nil_on_output_with_deprecated_required.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class DisallowNilOnOutputWithDeprecatedRequired < Actor
+  output :name, required: true
+
+  input :test_without_output, default: false
+
+  def call
+    context.name = 'Jim' unless test_without_output
+  end
+end