+++ spec/examples/set_required_output.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class SetRequiredOutput < Actor
-  output :name, required: true
-
-  def call
-    context.name = 'Jim'
-  end
-end