+++ spec/examples/set_wrong_required_output.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class SetWrongRequiredOutput < Actor
-  output :name, required: true
+  output :name, allow_nil: false
 
   def call
     # Expected to fail since it is required and we don't output `name`.